generator/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
firebase-admin==7.0.0
openai==1.97.0
python-dotenv==1.1.1
numpy==1.24.3
pydantic==2.11.7 